ParkingApp/send-email/requirements.txt
functions-framework
google-cloud-storage
google-api-python-client
google-auth
google-auth-oauthlib
